Caption (title) is now always the deciding factor
We fully rely on caption to be the deciding factor when identifying
a window to restore. Previously, if a window caption was mostly matching
but the window dimentions matched better, this window would be selected.

Now we always select the window with best matching caption, and size is
the secondary deciding factor. This should increase the probability of
correctly restoring all windows.